Hotfix writer emails updates (#1141)
* Update writer validation email

* Allow update for unverified users and add make exam notification for team when writer account is validated

* Add accept and cancel article buttons

* Fix article totals

* Send article notification queue

* Writer validation page updates

* Add notifications for writer exam

* Add article in writing reminder emails

* Add survey page

* Add writer survey

* Add writer survey graphs for admin

* Add internal writer access to writer validation and notification

+++ app/Repositories/AccountRepository.php
@@ -36,7 +36,7 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
 
         $response['success'] = false;
         $input_2 = collect($data)->only('update_method_payment_type')->toArray();
-        $input = collect($data)->except('company_type', 'user')->toArray();
+        $input = collect($data)->except('company_type', 'user', 'isVerified')->toArray();
 
         $input['is_freelance'] = $data['company_type'] == 'Freelancer' ? 1:0;
         $input['is_show_price_basis'] = $data['is_show_price_basis'] == 'yes' ? 1:0;
@@ -69,7 +69,7 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
             if(isset($input['affiliate'])) {
                 $account->update(['affiliate_id' => $input['affiliate']]);
             }
-            
+
         }
 
         if (isset($input['company_url']) && !empty($input['company_url'])) {
@@ -81,55 +81,75 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
         $user = User::where('email', $account->email)->first();
 
         if (isset($input['password']) && $input['password'] != '') {
-            $user->update(['password' => $input['password']]);
+            if ($user) {
+                $user->update(['password' => $input['password']]);
+            }
         }
 
         if (isset($input['credit_auth']) && $input['credit_auth'] != '') {
-            $user->update(['credit_auth' => $input['credit_auth']]);
+            if ($user) {
+                $user->update(['credit_auth' => $input['credit_auth']]);
 
-            $emails = $user->subBuyers()->pluck('email');
+                $emails = $user->subBuyers()->pluck('email');
 
-            User::whereIn('email', $emails)->update([
-                'credit_auth' => $input['credit_auth']
-            ]);
+                User::whereIn('email', $emails)->update([
+                    'credit_auth' => $input['credit_auth']
+                ]);
 
-            Registration::whereIn('email', $emails)->update([
-                'credit_auth' => $input['credit_auth']
-            ]);
+                Registration::whereIn('email', $emails)->update([
+                    'credit_auth' => $input['credit_auth']
+                ]);
+            }
         }
 
         if (isset($input['email']) && $input['email'] != '') {
-            $user->update(['email' => $input['email']]);
+            if ($user) {
+                $user->update(['email' => $input['email']]);
+            }
         }
 
         if (isset($input['username']) && $input['username'] != '') {
-            $user->update(['username' => $input['username']]);
+            if ($user) {
+                $user->update(['username' => $input['username']]);
+            }
         }
 
         if (isset($input['status']) && $input['status'] != '') {
-            $user->update(['status' => $input['status']]);
+            if ($user) {
+                $user->update(['status' => $input['status']]);
+            }
         }
 
         if (isset($input['account_validation']) && $input['account_validation'] != '') {
             if ($input['account_validation'] == 'invalid' || $input['account_validation'] == 'Invalid') {
-                $user->update(['status' => 'inactive']);
+                if ($user) {
+                    $user->update(['status' => 'inactive']);
+                }
 
                 $input['status'] = 'inactive';
             }
         }
 
         if (isset($input['name']) && $input['name'] != '') {
-            $user->update(['name' => $input['name']]);
+            if ($user) {
+                $user->update(['name' => $input['name']]);
+            }
         }
 
         if (isset($input['skype']) && $input['skype'] != '') {
-            $user->update(['skype' => $input['skype']]);
+            if ($user) {
+                $user->update(['skype' => $input['skype']]);
+            }
         } else {
-            $user->update(['skype' => 'none']);
+            if ($user) {
+                $user->update(['skype' => 'none']);
+            }
         }
 
         if (isset($input['id_payment_type']) && $input['id_payment_type'] != '') {
-            $user->update(['id_payment_type' => $input['id_payment_type']]);
+            if ($user) {
+                $user->update(['id_payment_type' => $input['id_payment_type']]);
+            }
         }
 
         // change user role_id according to type
@@ -142,7 +162,9 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
                 'Affiliate' => 11
             ];
 
-            $user->update(['role_id' => $user_roles[$input['type']]]);
+            if ($user) {
+                $user->update(['role_id' => $user_roles[$input['type']]]);
+            }
         }
 
         // ---------------------------------------------------
@@ -151,7 +173,11 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
 
         if (isset($input['account_validation']) && $input['account_validation'] === 'valid') {
             if ($account->email_via !== 'account_validated' && $account->type !== 'Affiliate') {
-                event(new UserValidateEvent($input, $user));
+
+                if ($user) {
+                    event(new UserValidateEvent($input, $user));
+                }
+
                 $input['email_via'] = 'account_validated';
                 $input['validation_reminded_at'] = null;
                 $input['reminded_days'] = null;
@@ -172,7 +198,9 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
 
         if (isset($input['team_in_charge']) && !empty($input['team_in_charge'])) {
             if ($account->team_in_charge !== $input['team_in_charge']) {
-                event(new TeamInChargeUpdatedEvent($input['team_in_charge'], $user));
+                if ($user) {
+                    event(new TeamInChargeUpdatedEvent($input['team_in_charge'], $user));
+                }
             }
         }
 
@@ -182,7 +210,11 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
 
         if (isset($input['status']) && $input['status'] === 'inactive') {
             if ($account->email_via_others !== 'account_invalidated') {
-                event(new UserUnvalidatedEvent($input, $user));
+
+                if ($user) {
+                    event(new UserUnvalidatedEvent($input, $user));
+                }
+
                 $input['email_via_others'] = 'account_invalidated';
                 $input['validation_reminded_at'] = null;
                 $input['email_via'] = null;
@@ -205,28 +237,30 @@ class AccountRepository extends BaseRepository implements AccountRepositoryInter
 
         // ---------------------------------------------------
 
-        $users_payment_type = UsersPaymentType::where('user_id', $user->id);
-        $users_payment_type->delete();
+        if ($user) {
+            $users_payment_type = UsersPaymentType::where('user_id', $user->id);
+            $users_payment_type->delete();
 
-        // Insert users payments types
-        $insert_input_users_payment_type = [];
-        if(is_array($input_2['update_method_payment_type'])) {
-            foreach($input_2['update_method_payment_type'] as $key => $types) {
-                if($types != '') {
-                    array_push($insert_input_users_payment_type, [
-                        'user_id' => $user->id,
-                        'payment_id' => $key,
-                        'account' => $types,
-                        'is_default' => $key == $input['id_payment_type'] ? 1:0,
-                        'created_at' => Carbon::now(),
-                        'updated_at' => Carbon::now()
-                    ]);
+            // Insert users payments types
+            $insert_input_users_payment_type = [];
+            if(is_array($input_2['update_method_payment_type'])) {
+                foreach($input_2['update_method_payment_type'] as $key => $types) {
+                    if($types != '') {
+                        array_push($insert_input_users_payment_type, [
+                            'user_id' => $user->id,
+                            'payment_id' => $key,
+                            'account' => $types,
+                            'is_default' => $key == $input['id_payment_type'] ? 1:0,
+                            'created_at' => Carbon::now(),
+                            'updated_at' => Carbon::now()
+                        ]);
+                    }
                 }
             }
-        }
 
-        if (count($insert_input_users_payment_type)) {
-            UsersPaymentType::insert($insert_input_users_payment_type);
+            if (count($insert_input_users_payment_type)) {
+                UsersPaymentType::insert($insert_input_users_payment_type);
+            }
         }
 
         $response['success'] = true;
